Refactor codebase to follow YAGNI, KISS, and DRY rules, Ruby style an…
