--- corrector.rb
@@ -1,5 +1,0 @@
-class Corrector
-  def correct_name(name)
-    name.capitalze[0, 10].strip
-  end
-end
