+++ library/book.rb
