+++ library/capitalizedecorator.rb
