+++ library/create_book.rb
@@ -1,0 +1,13 @@
+class CreateBook
+  def create_book(books)
+    @books = books
+    print 'Title: '
+    title = gets.chomp.strip
+
+    print 'Author: '
+    author = gets.chomp.strip
+    @books << Book.new(title, author)
+
+    puts 'Book created successfully'
+  end
+end
