+++ library/decorator.rb
