+++ library/list_people.rb
@@ -1,0 +1,8 @@
+class ListPeople
+  def list_people(people)
+    @people = people
+    @people.each_with_index do |person, index|
+      puts "#{index + 1}) [#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
+    end
+  end
+end
