+++ library/rental.rb
