+++ library/trimmerdecorator.rb
